rearrange file layouts.
