+++ clmctfy/clmctfy_container_api_raw.cc
@@ -1,0 +1,3 @@
+#include "clmctfy_container_api_raw.h"
+
+#define STATUS_OK UTIL__ERROR__CODE__OK
