+++ clmctfy/clmctfy_container_api_struct.h
@@ -1,0 +1,10 @@
+#ifndef LMCTFY_CLMCTFY_CLMCTFY_CONTAINER_API_STRUCT_H_
+#define LMCTFY_CLMCTFY_CLMCTFY_CONTAINER_API_STRUCT_H_
+
+#include "lmctfy.h"
+
+struct container_api {
+  ::containers::lmctfy::ContainerApi *container_api_;
+};
+
+#endif // LMCTFY_CLMCTFY_CLMCTFY_CONTAINER_API_STRUCT_H_
