+++ clmctfy/clmctfy_status_internal.cc
@@ -1,11 +1,9 @@
-#include "clmctfy_internal.h"
+#include "clmctfy_status_internal.h"
 
 #include <string.h> // to use strdup
 #include <stdarg.h>
 
 #include "util/task/status.h"
-
-#include "clmctfy.h"
 
 using ::util::Status;
